Fixed tests duplications; refs #36843 #37372
