+++ 202/test/AxeptiocookiesAddon/Hook/CommonHookTest.php
@@ -19,54 +19,11 @@
 
 namespace AxeptiocookiesAddon\Hook;
 
-use AxeptiocookiesAddon\Entity\AxeptioConfiguration;
-use AxeptiocookiesAddon\Entity\AxeptioModuleConfiguration;
-use AxeptiocookiesAddon\Model\CreateConfigurationModel;
-use AxeptiocookiesAddon\Service\ConfigurationService;
+use AxeptiocookiesAddon\AxeptioBaseTestCase;
 use AxeptiocookiesAddon\Service\HookService;
-use AxeptiocookiesAddon\Service\ModuleService;
-use AxeptiocookiesAddon\Utils\ServiceContainer;
-use FrontController;
-use Module;
-use PHPUnit\Framework\TestCase;
 
-class CommonHookTest extends TestCase
+class CommonHookTest extends AxeptioBaseTestCase
 {
-    /**
-     * @var HookDispatcher
-     */
-    protected $hookDispatcher;
-
-    /**
-     * @var ConfigurationService
-     */
-    protected $configurationService;
-
-    /**
-     * @var ModuleService
-     */
-    protected $moduleService;
-
-    /**
-     * @var Module
-     */
-    protected $module;
-
-    protected function setUp()
-    {
-        parent::setUp();
-        $this->module = ServiceContainer::getInstance()->get('axeptiocookies.module');
-        $this->hookDispatcher = new HookDispatcher($this->module);
-        $this->configurationService = ServiceContainer::getInstance()->get(ConfigurationService::class);
-        $this->moduleService = ServiceContainer::getInstance()->get(ModuleService::class);
-    }
-
-    public static function setUpBeforeClass()
-    {
-        parent::setUpBeforeClass();
-        self::truncateTables();
-    }
-
     public function testActionDispatcherBeforeFrontEmpty()
     {
         $result = $this->hookDispatcher->dispatch(
@@ -120,35 +77,5 @@ class CommonHookTest extends TestCase
         $this->createConfigurationFixtures();
         $result = $this->hookDispatcher->dispatch('displayFooter');
         $this->assertNotEmpty($result);
-    }
-
-    private function createConfigurationFixtures()
-    {
-        $hook = [
-            'id_hook' => \Hook::getIdByName('displayFooterBefore'),
-            'module' => 'ps_emailsubscription',
-            'id_module' => \Module::getModuleIdByName('ps_emailsubscription'),
-        ];
-        $createConfigurationModel = (new CreateConfigurationModel())
-            ->setIdProject(getenv('TEST_ID_PROJECT'))
-            ->setIdConfiguration(getenv('TEST_ID_CONFIGURATION'))
-            ->setIdLanguage(1)
-            ->setIdShops([1]);
-
-        $createConfigurationId = $this->configurationService->createConfiguration($createConfigurationModel);
-        $this->moduleService->associateToModules($createConfigurationId, [
-            'ps_emailsubscription',
-        ]);
-
-        return [
-            'hooks' => [$hook],
-            'id' => $createConfigurationId,
-        ];
-    }
-
-    private static function truncateTables()
-    {
-        \Db::getInstance()->delete(AxeptioConfiguration::$definition['table'], 1);
-        \Db::getInstance()->delete(AxeptioModuleConfiguration::$definition['table'], 1);
     }
 }
